apps/ai-api/pyproject.toml
[project]
name = "ai-api"
version = "0.1.0"
description = ""
authors = [
    {name = "JungJoonHo",email = "wnsgh6749@naver.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = []

[tool.poetry]
packages = [{include = "ai_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^0.3.27"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
openai-whisper = "^20231117"
torch = "^2.0.0"
torchaudio = "^2.0.0"
pydantic = "^2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
